requirements.txt
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==5.1
attrs==25.1.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.6
django-ckeditor==6.7.2
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-rest==0.8.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.9
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pymessenger==0.0.7.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.9.0
